simple-vision-rag/celery_worker/pyproject.toml
[project]
name = "celery-worker"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.5.3",
    "loguru>=0.7.3",
    "pdf2image>=1.17.0",
    "pillow>=11.3.0",
    "pydantic-settings>=2.11.0",
    "qdrant-client>=1.15.1",
    "redis>=6.4.0",
    "requests>=2.32.5",
]
